pyproject.toml
[project]
name = "pydwarf"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyelftools>=0.32",
    "tabulate>=0.10.0",
]
